Code cleanup and implemented more pictures for frontend rendering
